+++ test/jest.config.cjs
@@ -2,7 +2,10 @@ module.exports = {
   preset: 'ts-jest',
   rootDir: '..',
   testEnvironment: 'node',
-  transform: { '^.+\\.tsx?$': ['ts-jest', { tsconfig: '<rootDir>/test/tsconfig.json' }] },
+  extensionsToTreatAsEsm: ['.ts'],
+  transform: {
+    '^.+\\.tsx?$': ['ts-jest', { useESM: true, tsconfig: '<rootDir>/test/tsconfig.json' }],
+  },
   roots: ['<rootDir>/test'],
   testMatch: ['**/*.spec.ts'],
   collectCoverageFrom: ['src/**/*.ts'],